This produces a working (tested) image.
I think the main issue was that the image was being written as qcow2,
and this forces a raw image to be output.  Also added some extra packages
that needed to be installed, and added jenkins user to sudoers file.

Finally added a very brief readme with instructions on what is required
to get this to build on a centos 7 machine.